Rename folder 'fuctions' to 'Function'
