+++ ManagementPages/Model/Interfaces/IPostModel.cs
@@ -1,5 +1,5 @@
 ﻿using System.Threading.Tasks;
-using ManagementPages.Functions;
+using ManagementPages.Function;
 
 namespace ManagementPages.Model
 {
