+++ ManagementPages/Model/Models/InformationBoardModel.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using ManagementPages.Functions;
+using ManagementPages.Function;
 
 namespace ManagementPages.Model
 {
