+++ ManagementPages/Model/Models/LicenseModel.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using ManagementPages.Functions;
+using ManagementPages.Function;
 
 namespace ManagementPages.Model
 {
