+++ ManagementPages/UI/Components/Header.razor
@@ -1,5 +1,4 @@
-﻿@using ManagementPages.Functions
-@using ManagementPages.Model
+﻿@using ManagementPages.Model
 @using ManagementPages.UI.Components.Modals
 
 <Help @ref="Help" />
