+++ ManagementPages/UI/Components/InformationBoardOverview.razor
@@ -1,4 +1,3 @@
-@using ManagementPages.Functions
 @using ManagementPages.Model
 @using ManagementPages.UI.Components.Modals
