+++ ManagementPages/UI/Components/Modals/EditCategories.razor
@@ -1,6 +1,6 @@
-﻿@using ManagementPages.Functions
-@using ManagementPages.Model
+﻿@using ManagementPages.Model
 @using Syncfusion.Blazor.DropDowns
+@using ManagementPages.Function
 @inject IDbService _dbService
 
 <ManagementPages.UI.Components.Modals.DeleteCategory @ref="DeleteCategory" OnChangesMade="OnChangesMadeInChildComponent"/>
