+++ ManagementPages/UI/Components/Modals/EditInfoboard.razor
@@ -1,5 +1,5 @@
-﻿@using ManagementPages.Functions
-@using ManagementPages.Model
+﻿@using ManagementPages.Model
+@using ManagementPages.Function
 @inject IDbService _dbService
