+++ ManagementPages/UI/Components/Modals/NewCategory.razor
@@ -1,6 +1,6 @@
-@using ManagementPages.Functions
 @using ManagementPages.Model
 @using Syncfusion.Blazor.DropDowns
+@using ManagementPages.Function
 @inject IDbService _dbService
 
 <div class="modal @ModalClass" tabindex="-1" role="dialog" style="display: @ModalDisplay">
