+++ ManagementPages/UI/Components/Modals/NewPost.razor
@@ -1,5 +1,5 @@
-﻿@using ManagementPages.Functions
-@using ManagementPages.Model
+﻿@using ManagementPages.Model
+@using ManagementPages.Function
 @inject IDbService _dbService
 
 <div class="modal @ModalClass" tabindex="-1" role="dialog" style="display: @ModalDisplay">
