+++ ManagementPages/UI/Components/PostOverview.razor
@@ -1,6 +1,6 @@
-@using ManagementPages.Functions
 @using ManagementPages.Model
 @using ManagementPages.UI.Components.Modals
+@using ManagementPages.Function
 
 @inject IDbService _dbService
