+++ ManagementPages/UI/Pages/Admin.razor
@@ -1,7 +1,7 @@
 ﻿@page "/admin"
-@using ManagementPages.Functions
 @using ManagementPages.Model
 @using ManagementPages.UI.Components
+@using ManagementPages.Function
 
 @inject IDbService _dbService
