+++ ManagementPages/UI/Pages/Index.razor
@@ -1,7 +1,6 @@
 ﻿@page "/"
-
-@using ManagementPages.Functions
 @using ManagementPages.Model
+@using ManagementPages.Function
 @inject IDbService dbService
 @inject ILicenseModel licenseViewModel
